registering picture fill scripts is often too late
While it's proper to use the wp_enqueue_scripts when caling wp_enqueue_script, it is not a good idea for using it for wp_register_script.

For instance, the way it is written, there is no registration of the picturefill script on admin pages. Also, because of the triggering ordering, is is hairy to enqueue scripts in something that requires a wp_iframe.

Moving the registration to init.

Signed-off-by: terry chay <tychay@php.net>