refactored picture_wp_clear_options()
picture_wp_clear_options() and picturefill_wp_clear_transients() are redundant.

I spent some time thinking about it, and clear_transients() is called when updating javascript while clear_options() is called when deactivating the plugin, so it is theoretically possible that they should remain different functions.

However, if you are deactivating, then clearly you want to always clear the transients so, the former should call the latter instead of do the work of the latter.

Signed-off-by: terry chay <tychay@php.net>