my attempt to fix Issue #57
https://github.com/kylereicks/picturefill.js.wp/issues/57

Clearly the code I'm patching is wrong, but I'm not sure this will fix it. My thinking is that the intention of this section of code is to get the expected width/height tags of an image if it was not provided. But personally, I feel if it's not provided the intent isn't necessarily to fix it. :-(

Signed-off-by: terry chay <tychay@php.net>